+++ tests/cli.test.ts
@@ -110,7 +110,7 @@ describe("CLI", () => {
     const r = await cli(["--load", "tests/data", "--environment", "dev", "--", nodeCommand("console.log(process.env.DEV);")]);
     expect(r.success).toBeTruthy();
     expect(r.stdout).toContain("true");
-    });
+  });
 
   test("should load env-folder with environment and handle env-vars", async () => {
     const r = await cli(["--var", "DEV=false", "--load", "tests/data", "--environment", "dev", "--", nodeCommand("console.log(process.env.DEV);")]);