+++ tests/index.test.ts
@@ -1,15 +1,15 @@
-import { load, parse, parseString, values } from "../src";
+import {load, parse, parseString, values} from "../src";
 
 describe("dotenv-ng", () => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   let prevEnv: any;
 
   beforeEach(() => {
-    prevEnv = { ...process.env };
+    prevEnv = {...process.env};
   });
 
   afterEach(() => {
-    process.env = { ...prevEnv };
+    process.env = {...prevEnv};
   });
 
   test("should parse default env file with defaults", () => {
@@ -33,12 +33,47 @@ describe("dotenv-ng", () => {
         OTHER_NULL_VARIABLE: null,
         UNDEFINED_VARIABLE: undefined,
         OTHER_UNDEFINED_VARIABLE: undefined,
-        LITERAL_NULL_VARIABLE: "null"
+        LITERAL_NULL_VARIABLE: "null",
       },
       errors: [
-        { line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED" },
-        { line: 25, error: "MISSING_KEY", data: "=\"this as well.\"" }
-      ]
+        {line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED"},
+        {line: 25, error: "MISSING_KEY", data: "=\"this as well.\""}
+      ],
+      optional: []
+    })
+  });
+
+  test("should parse default env file with normalized names", () => {
+    process.env.SIMPLE_STRING_VARIABLE = "NOT TO BE OVERWRITTEN";
+    expect(parse({
+      normalize: true
+    })).toEqual({
+      data: {
+        EXPORTED_VARIABLE: "simple value",
+        OTHER_EXPORTED_VARIABLE: 12123.13,
+        SIMPLE_STRING_VARIABLE: "hello world",
+        OTHER_STRING_VARIABLE: "hello world",
+        SIMPLE_NUMBER_VARIABLE: 1,
+        SIMPLE_BOOLEAN_VARIABLE: false,
+        OTHER_BOOLEAN_VARIABLE: true,
+        OTHER_CASE_BOOLEAN_VARIABLE: true,
+        INTERPOLATED_VARIABLE: "this is also NOT TO BE OVERWRITTEN",
+        OTHER_BUT_NOT_INTERPOLATED: "this won't work $SIMPLE_STRING_VARIABLE (for now)",
+        INTERPOLATED_WITH_SYSVARS: `system temp: ${process.env.HOME}`,
+        "this is also an environment variable": "with this value",
+        THIS_IS_ALSO_AN_ENVIRONMENT_VARIABLE: "with this value",
+        EMPTY_VARIABLE: "",
+        NULL_VARIABLE: null,
+        OTHER_NULL_VARIABLE: null,
+        UNDEFINED_VARIABLE: undefined,
+        OTHER_UNDEFINED_VARIABLE: undefined,
+        LITERAL_NULL_VARIABLE: "null",
+      },
+      errors: [
+        {line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED"},
+        {line: 25, error: "MISSING_KEY", data: "=\"this as well.\""}
+      ],
+      optional: []
     })
   });
 
@@ -65,12 +100,13 @@ describe("dotenv-ng", () => {
         OTHER_NULL_VARIABLE: "NULL",
         UNDEFINED_VARIABLE: "undefined",
         OTHER_UNDEFINED_VARIABLE: "UNDEFINED",
-        LITERAL_NULL_VARIABLE: "null"
+        LITERAL_NULL_VARIABLE: "null",
       },
       errors: [
-        { line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED" },
-        { line: 25, error: "MISSING_KEY", data: "=\"this as well.\"" }
-      ]
+        {line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED"},
+        {line: 25, error: "MISSING_KEY", data: "=\"this as well.\""}
+      ],
+      optional: []
     })
   });
 
@@ -94,12 +130,14 @@ describe("dotenv-ng", () => {
         OTHER_NULL_VARIABLE: null,
         UNDEFINED_VARIABLE: undefined,
         OTHER_UNDEFINED_VARIABLE: undefined,
-        LITERAL_NULL_VARIABLE: "null"
+        LITERAL_NULL_VARIABLE: "null",
+        FILE: ".env",
       },
       errors: [
-        { line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED" },
-        { line: 25, error: "MISSING_KEY", data: "=\"this as well.\"" }
-      ]
+        {line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED"},
+        {line: 25, error: "MISSING_KEY", data: "=\"this as well.\""}
+      ],
+      optional: []
     })
   });
 
@@ -127,10 +165,12 @@ describe("dotenv-ng", () => {
         OTHER_UNDEFINED_VARIABLE: undefined,
         LITERAL_NULL_VARIABLE: "null",
         THIS_WILL_BE_IGNORED: "",
+        FILE: ".env",
       },
       errors: [
-        { line: 25, error: "MISSING_KEY", data: "=\"this as well.\"" }
-      ]
+        {line: 25, error: "MISSING_KEY", data: "=\"this as well.\""}
+      ],
+      optional: []
     })
   });
 
@@ -155,13 +195,15 @@ describe("dotenv-ng", () => {
         OTHER_NULL_VARIABLE: null,
         UNDEFINED_VARIABLE: undefined,
         OTHER_UNDEFINED_VARIABLE: undefined,
-        LITERAL_NULL_VARIABLE: "null"
+        LITERAL_NULL_VARIABLE: "null",
+        FILE: ".env",
       },
       errors: [
-        { line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED" },
-        { line: 25, error: "MISSING_KEY", data: "=\"this as well.\"" },
-        { line: 27, error: "EMPTY_VARIABLE", data: "EMPTY_VARIABLE=" }
-      ]
+        {line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED"},
+        {line: 25, error: "MISSING_KEY", data: "=\"this as well.\""},
+        {line: 27, error: "EMPTY_VARIABLE", data: "EMPTY_VARIABLE="}
+      ],
+      optional: []
     })
   });
 
@@ -187,12 +229,14 @@ describe("dotenv-ng", () => {
         OTHER_NULL_VARIABLE: "NULL",
         UNDEFINED_VARIABLE: undefined,
         OTHER_UNDEFINED_VARIABLE: "UNDEFINED",
-        LITERAL_NULL_VARIABLE: "null"
+        LITERAL_NULL_VARIABLE: "null",
+        FILE: ".env",
       },
       errors: [
-        { line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED" },
-        { line: 25, error: "MISSING_KEY", data: "=\"this as well.\"" }
-      ]
+        {line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED"},
+        {line: 25, error: "MISSING_KEY", data: "=\"this as well.\""}
+      ],
+      optional: []
     })
   });
 
@@ -218,12 +262,14 @@ describe("dotenv-ng", () => {
         OTHER_NULL_VARIABLE: "NULL",
         UNDEFINED_VARIABLE: "undefined",
         OTHER_UNDEFINED_VARIABLE: "UNDEFINED",
-        LITERAL_NULL_VARIABLE: "null"
+        LITERAL_NULL_VARIABLE: "null",
+        FILE: ".env",
       },
       errors: [
-        { line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED" },
-        { line: 25, error: "MISSING_KEY", data: "=\"this as well.\"" }
-      ]
+        {line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED"},
+        {line: 25, error: "MISSING_KEY", data: "=\"this as well.\""}
+      ],
+      optional: []
     })
   });
 
@@ -250,12 +296,49 @@ describe("dotenv-ng", () => {
         OTHER_NULL_VARIABLE: "NULL",
         UNDEFINED_VARIABLE: "undefined",
         OTHER_UNDEFINED_VARIABLE: "UNDEFINED",
-        LITERAL_NULL_VARIABLE: "null"
+        LITERAL_NULL_VARIABLE: "null",
+        FILE: ".env",
       },
       errors: [
-        { line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED" },
-        { line: 25, error: "MISSING_KEY", data: "=\"this as well.\"" }
-      ]
+        {line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED"},
+        {line: 25, error: "MISSING_KEY", data: "=\"this as well.\""}
+      ],
+      optional: []
+    })
+  });
+
+  test("should parse env file with defaults", () => {
+    expect(parse("tests/data/.env.default", {
+      parseLiterals: false,
+      parseNumbers: false,
+    })).toEqual({
+      data: {
+        EXPORTED_VARIABLE: "simple value",
+        OTHER_EXPORTED_VARIABLE: "12_123.13",
+        SIMPLE_STRING_VARIABLE: "hello world",
+        OTHER_STRING_VARIABLE: "hello world",
+        SIMPLE_NUMBER_VARIABLE: "1",
+        SIMPLE_BOOLEAN_VARIABLE: "false",
+        OTHER_BOOLEAN_VARIABLE: "true",
+        OTHER_CASE_BOOLEAN_VARIABLE: "TRUE",
+        INTERPOLATED_VARIABLE: "this is also hello world",
+        OTHER_BUT_NOT_INTERPOLATED: "this won't work $SIMPLE_STRING_VARIABLE (for now)",
+        INTERPOLATED_WITH_SYSVARS: `system temp: ${process.env.HOME}`,
+        "this is also an environment variable": "with this value",
+        EMPTY_VARIABLE: "",
+        NULL_VARIABLE: "null",
+        OTHER_NULL_VARIABLE: "NULL",
+        UNDEFINED_VARIABLE: "undefined",
+        OTHER_UNDEFINED_VARIABLE: "UNDEFINED",
+        LITERAL_NULL_VARIABLE: "null",
+        FILE: ".env.default",
+        DOES_NOT_EXIST_YET: "DEFAULT",
+      },
+      errors: [
+        {line: 24, error: "ORPHAN_KEY", data: "THIS_WILL_BE_IGNORED"},
+        {line: 25, error: "MISSING_KEY", data: "=\"this as well.\""}
+      ],
+      optional: ["DOES_NOT_EXIST_YET", "FILE"]
     })
   });
 
@@ -264,7 +347,8 @@ describe("dotenv-ng", () => {
       data: {
         HELLO: "${HELLO} WORLD"
       },
-      errors: []
+      errors: [],
+      optional: []
     });
   });
 
@@ -287,7 +371,8 @@ describe("dotenv-ng", () => {
       OTHER_NULL_VARIABLE: null,
       UNDEFINED_VARIABLE: undefined,
       OTHER_UNDEFINED_VARIABLE: undefined,
-      LITERAL_NULL_VARIABLE: "null"
+      LITERAL_NULL_VARIABLE: "null",
+      FILE: ".env",
     })
   });
 
@@ -313,7 +398,8 @@ describe("dotenv-ng", () => {
       OTHER_NULL_VARIABLE: "NULL",
       UNDEFINED_VARIABLE: "undefined",
       OTHER_UNDEFINED_VARIABLE: "UNDEFINED",
-      LITERAL_NULL_VARIABLE: "null"
+      LITERAL_NULL_VARIABLE: "null",
+      FILE: ".env",
     })
   });
 
@@ -337,7 +423,8 @@ describe("dotenv-ng", () => {
       OTHER_NULL_VARIABLE: null,
       UNDEFINED_VARIABLE: undefined,
       OTHER_UNDEFINED_VARIABLE: undefined,
-      LITERAL_NULL_VARIABLE: "null"
+      LITERAL_NULL_VARIABLE: "null",
+      FILE: ".env.local",
     })
   });
 
@@ -364,7 +451,8 @@ describe("dotenv-ng", () => {
       OTHER_NULL_VARIABLE: null,
       UNDEFINED_VARIABLE: undefined,
       OTHER_UNDEFINED_VARIABLE: undefined,
-      LITERAL_NULL_VARIABLE: "null"
+      LITERAL_NULL_VARIABLE: "null",
+      FILE: ".env.local",
     })
   });
 
@@ -390,7 +478,8 @@ describe("dotenv-ng", () => {
       OTHER_NULL_VARIABLE: null,
       UNDEFINED_VARIABLE: undefined,
       OTHER_UNDEFINED_VARIABLE: undefined,
-      LITERAL_NULL_VARIABLE: "null"
+      LITERAL_NULL_VARIABLE: "null",
+      FILE: ".env.local",
     })
   });
 
@@ -425,10 +514,10 @@ describe("dotenv-ng", () => {
       UNDEFINED_VARIABLE: undefined,
       OTHER_UNDEFINED_VARIABLE: undefined,
       LITERAL_NULL_VARIABLE: "null",
+      FILE: ".env.local",
       ...prevEnv,
     })
   });
-
 
   test("should load folder with overwriting process.env", () => {
     process.env.SIMPLE_STRING_VARIABLE = "TO BE OVERWRITTEN";
@@ -458,6 +547,39 @@ describe("dotenv-ng", () => {
       UNDEFINED_VARIABLE: undefined,
       OTHER_UNDEFINED_VARIABLE: undefined,
       LITERAL_NULL_VARIABLE: "null",
-    })
-  })
+      FILE: ".env.local",
+    });
+  });
+
+  test("should load folder with default values", () => {
+    process.env.SIMPLE_STRING_VARIABLE = "TO BE OVERWRITTEN";
+    load("tests/data", {
+      environment: "default",
+      overwriteExisting: true,
+    });
+    expect(process.env).toEqual({
+      ...prevEnv,
+      LOCAL: true,
+      EXPORTED_VARIABLE: "simple value",
+      OTHER_EXPORTED_VARIABLE: 12123.13,
+      SIMPLE_STRING_VARIABLE: "hello local",
+      OTHER_STRING_VARIABLE: "hello world",
+      SIMPLE_NUMBER_VARIABLE: 3,
+      SIMPLE_BOOLEAN_VARIABLE: false,
+      OTHER_BOOLEAN_VARIABLE: true,
+      OTHER_CASE_BOOLEAN_VARIABLE: true,
+      INTERPOLATED_VARIABLE: "this is also hello local",
+      OTHER_BUT_NOT_INTERPOLATED: "this won't work $SIMPLE_STRING_VARIABLE (for now)",
+      INTERPOLATED_WITH_SYSVARS: `system temp: ${prevEnv.HOME}`,
+      "this is also an environment variable": "with this value",
+      EMPTY_VARIABLE: "",
+      NULL_VARIABLE: null,
+      OTHER_NULL_VARIABLE: null,
+      UNDEFINED_VARIABLE: undefined,
+      OTHER_UNDEFINED_VARIABLE: undefined,
+      LITERAL_NULL_VARIABLE: "null",
+      DOES_NOT_EXIST_YET: "DEFAULT",
+      FILE: ".env.local",
+    });
+  });
 });