+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.2.0
+
+### Added
+
+- Added option to set default environment variables
+- Added support for variable name normalization
+
 ## 1.1.4
 
 ### Fixed